+++ README.md
@@ -60,9 +60,9 @@ yarn install
 
 | Package | Version | Dependencies | Dev dependencies | Peer dependencies | Changelog |
 | ------- | ------- | ------------ | ---------------- | ----------------- | :-------: |
-| [@ovh-ux/ng-ovh-uirouter-breadcrumb](https://github.com/ovh-ux/manager/tree/master/packages/components/ng-ovh-uirouter-breadcrumb) | [![npm version](https://badgen.net/npm/v/@ovh-ux/ng-ovh-uirouter-breadcrumb)](https://www.npmjs.com/package/@ovh-ux/ng-ovh-uirouter-breadcrumb) | [![Dependencies](https://badgen.net/david/dep/ovh-ux/manager/packages/components/ng-ovh-uirouter-breadcrumb)](https://npmjs.com/package/@ovh-ux/ng-ovh-uirouter-breadcrumb?activeTab=dependencies) | [![Dev Dependencies](https://badgen.net/david/dev/ovh-ux/manager/packages/components/ng-ovh-uirouter-breadcrumb)](https://npmjs.com/package/@ovh-ux/ng-ovh-uirouter-breadcrumb?activeTab=dependencies) | [![Peer Dependencies](https://badgen.net/david/peer/ovh-ux/manager/packages/components/ng-ovh-uirouter-breadcrumb)](https://npmjs.com/package/@ovh-ux/ng-ovh-uirouter-breadcrumb?activeTab=dependencies) | [:books:](https://github.com/ovh-ux/manager/blob/master/packages/components/ng-ovh-uirouter-breadcrumb/CHANGELOG.md) |
 | [@ovh-ux/ng-ovh-uirouter-layout](https://github.com/ovh-ux/manager/tree/master/packages/components/ng-ovh-uirouter-layout) | [![npm version](https://badgen.net/npm/v/@ovh-ux/ng-ovh-uirouter-layout)](https://www.npmjs.com/package/@ovh-ux/ng-ovh-uirouter-layout) | [![Dependencies](https://badgen.net/david/dep/ovh-ux/manager/packages/components/ng-ovh-uirouter-layout)](https://npmjs.com/package/@ovh-ux/ng-ovh-uirouter-layout?activeTab=dependencies) | [![Dev Dependencies](https://badgen.net/david/dev/ovh-ux/manager/packages/components/ng-ovh-uirouter-layout)](https://npmjs.com/package/@ovh-ux/ng-ovh-uirouter-layout?activeTab=dependencies) | [![Peer Dependencies](https://badgen.net/david/peer/ovh-ux/manager/packages/components/ng-ovh-uirouter-layout)](https://npmjs.com/package/@ovh-ux/ng-ovh-uirouter-layout?activeTab=dependencies) | [:books:](https://github.com/ovh-ux/manager/blob/master/packages/components/ng-ovh-uirouter-layout/CHANGELOG.md) |
 | [@ovh-ux/ng-ovh-uirouter-line-progress](https://github.com/ovh-ux/manager/tree/master/packages/components/ng-ovh-uirouter-line-progress) | [![npm version](https://badgen.net/npm/v/@ovh-ux/ng-ovh-uirouter-line-progress)](https://www.npmjs.com/package/@ovh-ux/ng-ovh-uirouter-line-progress) | [![Dependencies](https://badgen.net/david/dep/ovh-ux/manager/packages/components/ng-ovh-uirouter-line-progress)](https://npmjs.com/package/@ovh-ux/ng-ovh-uirouter-line-progress?activeTab=dependencies) | [![Dev Dependencies](https://badgen.net/david/dev/ovh-ux/manager/packages/components/ng-ovh-uirouter-line-progress)](https://npmjs.com/package/@ovh-ux/ng-ovh-uirouter-line-progress?activeTab=dependencies) | [![Peer Dependencies](https://badgen.net/david/peer/ovh-ux/manager/packages/components/ng-ovh-uirouter-line-progress)](https://npmjs.com/package/@ovh-ux/ng-ovh-uirouter-line-progress?activeTab=dependencies) | [:books:](https://github.com/ovh-ux/manager/blob/master/packages/components/ng-ovh-uirouter-line-progress/CHANGELOG.md) |
+| [@ovh-ux/ng-uirouter-breadcrumb](https://github.com/ovh-ux/manager/tree/master/packages/components/ng-uirouter-breadcrumb) | [![npm version](https://badgen.net/npm/v/@ovh-ux/ng-uirouter-breadcrumb)](https://www.npmjs.com/package/@ovh-ux/ng-uirouter-breadcrumb) | [![Dependencies](https://badgen.net/david/dep/ovh-ux/manager/packages/components/ng-uirouter-breadcrumb)](https://npmjs.com/package/@ovh-ux/ng-uirouter-breadcrumb?activeTab=dependencies) | [![Dev Dependencies](https://badgen.net/david/dev/ovh-ux/manager/packages/components/ng-uirouter-breadcrumb)](https://npmjs.com/package/@ovh-ux/ng-uirouter-breadcrumb?activeTab=dependencies) | [![Peer Dependencies](https://badgen.net/david/peer/ovh-ux/manager/packages/components/ng-uirouter-breadcrumb)](https://npmjs.com/package/@ovh-ux/ng-uirouter-breadcrumb?activeTab=dependencies) | [:books:](https://github.com/ovh-ux/manager/blob/master/packages/components/ng-uirouter-breadcrumb/CHANGELOG.md) |
 <!-- AUTO-GENERATED-CONTENT:END *-->
 
 ## Development
