--- packages/components/ng-ovh-uirouter-breadcrumb/src/ng-ovh-uirouter-breadcrumb.component.js
@@ -1,7 +1,0 @@
-import controller from './ng-ovh-uirouter-breadcrumb.controller';
-import template from './ng-ovh-uirouter-breadcrumb.html';
-
-export default {
-  controller,
-  template,
-};
