--- packages/components/ng-ovh-uirouter-breadcrumb/src/ng-ovh-uirouter-breadcrumb.controller.js
@@ -1,17 +1,0 @@
-export default class ngUirouterBreadcrumbController {
-  /* @ngInject */
-  constructor(
-    ngOvhUirouterBreadcrumbService,
-  ) {
-    this.ngOvhUirouterBreadcrumbService = ngOvhUirouterBreadcrumbService;
-
-    this.breadcrumb = [];
-  }
-
-  $onInit() {
-    this.ngOvhUirouterBreadcrumbService.subscribe((breadcrumb) => {
-      this.breadcrumb = breadcrumb;
-    });
-    this.breadcrumb = this.ngOvhUirouterBreadcrumbService.getBreadcrumb();
-  }
-}
