+++ packages/components/ng-uirouter-breadcrumb/README.md
@@ -1,0 +1,35 @@
+# ng-uirouter-breadcrumb
+
+
+[![Downloads](https://badgen.net/npm/dt/@ovh-ux/ng-uirouter-breadcrumb)](https://npmjs.com/package/@ovh-ux/ng-uirouter-breadcrumb) [![Dependencies](https://badgen.net/david/dep/ovh-ux/manager/packages/manager/modules/ng-uirouter-breadcrumb)](https://npmjs.com/package/@ovh-ux/ng-uirouter-breadcrumb?activeTab=dependencies) [![Dev Dependencies](https://badgen.net/david/dev/ovh-ux/manager/packages/manager/modules/ng-uirouter-breadcrumb)](https://npmjs.com/package/@ovh-ux/ng-uirouter-breadcrumb?activeTab=dependencies) [![Gitter](https://badgen.net/badge/gitter/ovh-ux/blue?icon=gitter)](https://gitter.im/ovh/ux)
+
+## Install
+
+```sh
+yarn add @ovh-ux/ng-uirouter-breadcrumb
+```
+## Usage
+
+```js
+import angular from 'angular';
+import ngUiRouterBreadcrumb from '@ovh-ux/ng-uirouter-breadcrumb';
+
+angular
+  .module('myApp', [
+    ngUiRouterBreadcrumb,
+  ]);
+```
+
+## Test
+
+```sh
+yarn test
+```
+
+## Contributing
+
+Always feel free to help out! Whether it's [filing bugs and feature requests](https://github.com/ovh-ux/manager/issues/new) or working on some of the [open issues](https://github.com/ovh-ux/manager/issues), our [contributing guide](CONTRIBUTING.md) will help get you started.
+
+## License
+
+[BSD-3-Clause](LICENSE) © OVH SAS
