+++ packages/components/ng-uirouter-breadcrumb/package.json
@@ -1,15 +1,17 @@
 {
-  "name": "@ovh-ux/ng-ovh-uirouter-breadcrumb",
+  "name": "@ovh-ux/ng-uirouter-breadcrumb",
   "version": "0.0.0",
   "description": "UiRouter breadcrumb",
   "keywords": [
-    "uirouter",
-    "ovh"
+    "angular",
+    "breadcrumb",
+    "ovh",
+    "uirouter"
   ],
   "repository": {
     "type": "git",
     "url": "git+https://github.com/ovh-ux/manager.git",
-    "directory": "packages/components/ng-ovh-uirouter-breadcrumb"
+    "directory": "packages/components/ng-uirouter-breadcrumb"
   },
   "license": "BSD-3-Clause",
   "author": "OVH SAS",
@@ -18,15 +20,15 @@
   ],
   "main": "./dist/cjs/index.js",
   "module": "./dist/esm/index.js",
-  "browser": "./dist/umd/ng-ovh-uirouter-breadcrumb.js",
+  "browser": "./dist/umd/ng-uirouter-breadcrumb.js",
   "scripts": {
     "build": "rollup -c --environment BUILD:production",
     "dev": "rollup -c --environment BUILD:development",
     "dev:watch": "yarn run dev --watch",
     "prepare": "yarn run build",
-    "start": "lerna exec --stream --scope='@ovh-ux/ng-ovh-uirouter-breadcrumb' --include-filtered-dependencies -- yarn run build",
-    "start:dev": "lerna exec --stream --scope='@ovh-ux/ng-ovh-uirouter-breadcrumb' --include-filtered-dependencies -- yarn run dev",
-    "start:watch": "lerna exec --stream --parallel --scope='@ovh-ux/ng-ovh-uirouter-breadcrumb' --include-filtered-dependencies -- yarn run dev:watch"
+    "start": "lerna exec --stream --scope='@ovh-ux/ng-uirouter-breadcrumb' --include-filtered-dependencies -- yarn run build",
+    "start:dev": "lerna exec --stream --scope='@ovh-ux/ng-uirouter-breadcrumb' --include-filtered-dependencies -- yarn run dev",
+    "start:watch": "lerna exec --stream --parallel --scope='@ovh-ux/ng-uirouter-breadcrumb' --include-filtered-dependencies -- yarn run dev:watch"
   },
   "devDependencies": {
     "@ovh-ux/component-rollup-config": "^5.0.0"
