+++ packages/components/ng-uirouter-breadcrumb/src/ng-uirouter-breadcrumb.controller.js
@@ -1,0 +1,17 @@
+export default class BreadcrumbController {
+  /* @ngInject */
+  constructor(
+    uirouterBreadcrumbService,
+  ) {
+    this.uirouterBreadcrumbService = uirouterBreadcrumbService;
+
+    this.breadcrumb = [];
+  }
+
+  $onInit() {
+    this.uirouterBreadcrumbService.subscribe((breadcrumb) => {
+      this.breadcrumb = breadcrumb;
+    });
+    this.breadcrumb = this.uirouterBreadcrumbService.getBreadcrumb();
+  }
+}
