+++ packages/components/ng-uirouter-breadcrumb/src/ng-uirouter-breadcrumb.less
@@ -1,4 +1,4 @@
-ng-ovh-uirouter-breadcrumb { /* stylelint-disable-line selector-type-no-unknown */
+uirouter-breadcrumb { /* stylelint-disable-line selector-type-no-unknown */
   ul {
     margin: 0;
     padding: 0;
