+++ packages/components/ng-uirouter-breadcrumb/src/ng-uirouter-breadcrumb.service.js
