+++ packages/manager/modules/pci/src/projects/project/project.html
@@ -8,7 +8,7 @@
         </div>
         <div class="h-100 position-relative col-md-10 p-0">
             <div class="position-absolute w-100 h-100 py-4 px-5" style="overflow: auto;">
-                <ng-ovh-uirouter-breadcrumb></ng-ovh-uirouter-breadcrumb>
+                <uirouter-breadcrumb></uirouter-breadcrumb>
                 <div class="mt-3" data-ui-view>
                     <div data-ng-if="$ctrl.project">
                         <oui-page-header heading="{{ ::$ctrl.project.description }}">
